claude-tooling/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
anthropic==0.21.1
python-dotenv==1.0.1
pydantic==2.6.3
starlette==0.36.3
jinja2==3.1.3
python-multipart==0.0.7
httpx==0.27.0
markupsafe==2.1.5 